Adding objects to scene1
Added a plane to represent the playing field, and a Boxbot with a rigidbody attached to it so we can start working on moving it around the playing field. Also added some basic materials to give the ground and boxbot some color. Finally I also added a scripts folder with an empty C# script entitled 'BoxbotController' that is attached to the game object Boxbot (hereafter referred to as BB) that should allow us to accept user input to mover BB around... once we figure out how to do so from a code standpoint.