Jgriego-Patch1-TapToMove
Added a script called TapToMove to the game object  'Model'; this allows the object to be moved by clicking with the mouse on a solid object in the game field. Credit for the code in TapToMove goes to Sujit Horakeri; http://www.thegamecontriver.com/2014/10/move-to-mouse-click-touch-position-unity.html

Also added a physics -> rigid body component to 'Model' that makes it obey physics; i.e. it will fall over when unbalanced and will not be able to pass through other physical objects in the game field.